Implement Get Class By ID gateway

+++ demos_back_golang/cmd/migrate/migrations/000002_create_classes.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS classes (
     id bigserial PRIMARY KEY,
     datetime timestamp(0) with time zone NOT NULL,
     location text NOT NULL,
-    price numeric(10,2) DEFAULT 0.00,
+    price INTEGER DEFAULT 0,
     trainer_id bigint NOT NULL REFERENCES users(id),
     created_at timestamp(0) with time zone NOT NULL DEFAULT NOW()
 );
